Fixes static initializers not being called on specific types
Previously it was assumed that instantiated_classes would be iterated through and then asserted that everything has a static initializer set up. This isn't the case anymore and this check (and insertion) is done when classes are marked as needed. This direct insert bypassed that functionality meaning certain classes were missing cproverNondetInitialize and clinit calls.